+++ storage_drivers/ontap/ontap_nas.go
@@ -163,6 +163,11 @@ func (d *NASStorageDriver) Create(name string, sizeBytes uint64, opts map[string
 		return err
 	}
 
+	snapshotReserve := api.NumericalValueNotSet
+	if snapshotPolicy == "none" {
+		snapshotReserve = 0
+	}
+
 	log.WithFields(log.Fields{
 		"name":            name,
 		"size":            size,
@@ -174,12 +179,13 @@ func (d *NASStorageDriver) Create(name string, sizeBytes uint64, opts map[string
 		"aggregate":       aggregate,
 		"securityStyle":   securityStyle,
 		"encryption":      encryption,
+		"snapshotReserve": snapshotReserve,
 	}).Debug("Creating Flexvol.")
 
 	// Create the volume
 	volCreateResponse, err := d.API.VolumeCreate(
 		name, aggregate, size, spaceReserve, snapshotPolicy,
-		unixPermissions, exportPolicy, securityStyle, encrypt)
+		unixPermissions, exportPolicy, securityStyle, encrypt, snapshotReserve)
 
 	if err = api.GetError(volCreateResponse, err); err != nil {
 		if zerr, ok := err.(api.ZapiError); ok {