+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - **Docker:** Fixed an issue where deleted qtrees could appear in Docker volume list.
 
 **Enhancements:**
+- Changed ONTAP drivers so that the snapshot reserve is set to zero when snapshotPolicy is "none" (Issue
 - **Kubernetes:** Updated etcd to v3.2.19 and client-go to v7.0.0.
 - **Kubernetes:** Added --previous switch to 'tridentctl logs' command.
 - **Kubernetes:** Added liveness probes to the trident-main and etcd containers.